pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "amati"
version = "0.1.0"
description = "Your project description"
authors = [{name = "Ben Alexander", email = "amati@ben-alexander.com"},]

dependencies = [
    "pydantic>=2.10.2",
    "pyyaml",
    "abnf>=2.3.0"
]
requires-python = ">=3.13"

[project.optional-dependencies]

dev = [
    "pytest>=7.0",
    "pytest-cov",
    "mutmut>=2.4.3",
    "hypothesis>=6.122.0",
    "autopep8",
    "pylint",
    "hammett"
]

build = [
    "requests"
]

[tool.setuptools]
packages = ["amati"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["."]


[tool.isort]
src_paths = ["amati", "tests"]
known_first_party=["amati"]
known_third_party=["pytest","pydantic","yaml","abnf"]
profile="black"